+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/chat/GroupChatImpl.kt
@@ -8,7 +8,6 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
 data class GroupChatImpl(
     override val id: ChatId,
     override val title: String? = null,
-    override val allMembersAreAdmins: Boolean,
     override val inviteLink: String? = null,
     override val chatPhoto: ChatPhoto? = null,
     override val pinnedMessage: RawMessage? = null,