+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changes according to [July 29, 2019 Telegram Bot API update](https://core.telegr
 
 * `Sticker` and `StickerSet` now have field `isAnimated`
 * `ChatPermissions` object was added, `GroupChat` interface got `permissions` field, request `SetChatPermissions` was added
+* `GroupChat` object now have no field `allMembersAreAdmins`
 
 ## 0.16.0 Bot API 4.3